aand then we draw a weird cross for no good reason
